moved everything so far into "builder"
