+++ typescript/packages/lookslike-high-level/test/builder/cell-proxy.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from "vitest";
-import { isCell } from "../src/framework/types.js";
-import { cell } from "../src/framework/cell-proxy.js";
+import { isCell } from "../../src/builder/types.js";
+import { cell } from "../../src/builder/cell-proxy.js";
 
 describe("cell function", () => {
   it("creates a cell proxy", () => {
